cre-plume: move timeseries/profile extractors to callback.py
