+++ thetis/__init__.py
@@ -5,6 +5,8 @@ import thetis.timeintegrator as timeintegrator  # NOQA
 import thetis.solver as solver  # NOQA
 import thetis.solver2d as solver2d  # NOQA
 from thetis.callback import DiagnosticCallback, DetectorsCallback  # NOQA
+from thetis.callback import TimeSeriesCallback2D, TimeSeriesCallback3D  # NOQA
+from thetis.callback import VerticalProfileCallback  # NOQA
 import thetis.limiter as limiter      # NOQA
 import thetis.interpolation as interpolation      # NOQA
 import thetis.coordsys as coordsys      # NOQA
